requirements.txt
pyserial==3.4
adafruit-io==2.4.0
